Use #message instead of #to_s
Exceptions raised by the google-adwords-api gem return the same thing for `e.class.to_s` and `e.to_s`. Switching to `message` here makes the output match what is shown in sidekiq's retries. I believe the convention is to implement a `message` method on exceptions, so I'm thinking this issue isn't specific to this gem.