Sidekiq 3 compatibility cleanup.
Sidekiq 3 just slightly changed its API. SortedSet is now more a base
class and we are interested in a concrete implementation which is now in
JobSet.

Simply reusing Sidekiq core class we reduce a burden to maintain our own
version of SortedSet/JobSet.

Conditional logic is needed to maintain compatibility with Sidekiq 2.x