Expect 3 calls for `options` on the manager when newer Sidekiq version
As of sidekiq v5.0.3, Sidekiq::Processor#initialize now calls `options`
on the given manager 3 times instead of 2, breaking the existing specs.
To fix, we simply tell the manager mock to expect the call 3 times when
the sidekiq version is new enough, and to expect 2 calls when the
sidekiq version is older.